Add product item trait
